auth/requirements.txt
alembic==1.8.1
bcrypt==4.0.1
click==7.1.2
fastapi==0.85.0
greenlet==2.0.1
h11==0.12.0
jsonschema==4.16.0
Mako==1.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==22.0
passlib==1.7.4
pexpect==4.8.0
pipenv==2022.12.19
pydantic==1.10.2
PyJWT==2.6.0
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
pytest==6.2.5
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.13.4
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1